zmq_py/requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mpmath==1.1.0
pylint==2.4.4
pyzmq==19.0.0
six==1.14.0
sympy==1.5.1
typed-ast==1.4.1
wrapt==1.11.2
